pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webseed"
version = "0.1.0"
description = "A modular web content scraping tool"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "requests>=2.26.0",
    "beautifulsoup4>=4.10.0",
    "selenium>=4.0.0",
    "webdriver-manager>=3.5.2",
    "pyyaml>=6.0",
    "html2text>=2020.1.16",
    "youtube_transcript_api>=0.6.0",
    "yt-dlp>=2023.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
    "isort>=5.9.1",
    "mypy>=0.910",
    "flake8>=3.9.2",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/webseed"
"Bug Tracker" = "https://github.com/yourusername/webseed/issues"

[project.scripts]
webseed = "webseed.cli:main"

[tool.setuptools]
packages = ["webseed"]

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
